requirements.txt
amqp==5.0.1
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.2.0
billiard==3.6.3.0
black==19.3b0
bleach==3.1.5
boto3==1.14.52
botocore==1.17.52
celery==5.0.0
certifi==2020.6.20
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
dj-rest-auth==1.1.1
Django==3.0
django-allauth==0.42.0
django-celery-email==3.0.0
django-cors-headers==3.5.0
django-countries==6.1.3
django-filter==2.3.0
django-guardian==2.3.0
django-jazzmin==2.2.7
django-redis==4.12.1
django-ses==1.0.3
django-storages==1.10
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
flake8==3.7.7
future==0.18.2
gunicorn==19.9.0
hurry.filesize==0.9
inflection==0.5.1
iniconfig==1.0.1
isort==4.3.20
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==5.0.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
oauthlib==3.1.0
pep8==1.7.1
py==1.9.0
pyflakes==2.1.1
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
pytest==6.0.1
python-dateutil==2.8.1
python-slugify==4.0.0
pytz==2020.1
readme-renderer==26.0
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
whitenoise==5.0
